Allow language kinds and loads to depend on config
Currently, gazelle has exactly one place where config is allowed to
affect per-build-file generated loads, and that's using `map_kind`. This
PR extends that mechanism to allow arbitrary languages to have arbitrary
config items which can be used to impact loaded symbols.

An example of this is shown in `tests/diff_mode_with_config` - each
BUILD file may define where its filegroup should be loaded from (or may
inherit that configuration from a parent BUILD file's configuration).

This is useful in practice for allowing individual consumers of a
gazelle plugin to configure their own customisations. For instance, in
Java a team may have a @Flaky annotation for test classes; this
mechanism allows a team to have config in their BUILD file to say "if
you see the `@Flaky` annotation on a test class, generate a
`flaky_java_test` for that class instead of a regular `java_test` class.

However, to be generally useful and extensible, this needs to be
configurable in the repo rather than the general purpose gazelle plugin
(and because monorepos may have separate teams working with different
conventions, in individual BUILD files in the repo), rather than
hard-coded in a particular plugin.

This was the cleanest, minimal modification I could find to enable this;
it allows for individual languages to have their `Load()` behaviour
depend on configuration.

There's a question about compatibility here; in an ideal world, we would
just modify `language.Language` so that `Kinds` and `Loads` took the
extra arguments, but that's a very breaking change. Currently, I've
implemented a middle-ground of adding a second interface, and sniffing
for its implementation when calling `Kinds` and Loads`, but this is
somewhat fragile. I'm not sure the best approach to making this kind of
change in the gazelle ecosystem. While a breaking change is somewhat
inconvenient, addressing it is trivial - just adding two unused
arguments to two methods. I think long-term, deprecation and migration
should be the answer, but I'm not sure how aggressive to be on the way
there.